Deleted bot_login.js
