+++ .gitignore
@@ -1,7 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-# bot_login
-bot_login.js
 bot_credentials.js
 
 # dependencies
